moved previews into scripts/
